+++ lang/en.php
@@ -22,3 +22,6 @@ $lang_str['form:check_lt_failed']="Value must be < %s";
 $lang_str['form:check_empty_or'] = "No parameters to 'or' operator";
 
 $lang_str['form_element:please_select'] = "-- please select --";
+
+$lang_str['form:check_all'] = "check all";
+$lang_str['form:uncheck_all'] = "uncheck all";